Update master to new versions

+++ model/src/main/scala/uk/gov/ons/addressIndex/model/server/response/address/AddressResponseHighlightHit.scala
@@ -1,6 +1,6 @@
 package uk.gov.ons.addressIndex.model.server.response.address
 
-import play.api.libs.json.{Format, Json};
+import play.api.libs.json.{Format, Json}
 
 case class AddressResponseHighlightHit(source: String,
                                        lang: String,
@@ -17,7 +17,7 @@ object AddressResponseHighlightHit {
     val searchField = hit._1.mkString
     val lang = if (searchField.contains("Welsh")) "W" else "E"
     val source = if (searchField.contains("Nag")) "L" else if (searchField.contains("Nisra")) "N" else "P"
-    val distinctHitCount = Math.round(highLightedText.mkString.replace(",","").split(" ").distinct.mkString.count(_ == '<') / 2)
+    val distinctHitCount = Math.round((highLightedText.mkString.replace(",","").split(" ").distinct.mkString.count(_ == '<') / 2).toFloat)
 
    Some(AddressResponseHighlightHit(source,lang,distinctHitCount,highLightedText))
   }
@@ -29,7 +29,7 @@ object AddressResponseHighlightHit {
     val searchField = testFieldName
     val lang = if (searchField.contains("elsh")) "W" else "E"
     val source = if (searchField.contains("Nag")) "L" else if (searchField.contains("Nisra")) "N" else "P"
-    val distinctHitCount = Math.round(highlightedTextOut.mkString.replace(",","").split(" ").distinct.mkString.count(_ == '<') / 2)
+    val distinctHitCount = Math.round((highlightedTextOut.mkString.replace(",","").split(" ").distinct.mkString.count(_ == '<') / 2).toFloat)
 
     Some(AddressResponseHighlightHit(source,lang,distinctHitCount,highlightedTextOut))
   }
